backend/requirements.txt
fastapi
uvicorn
transformers
torch
peft
sentence-transformers
faiss-cpu
pymupdf
opencv-python
pillow
numpy
requests
pandas
python-multipart